Arreglo del estado de zustand
