+++ .gitignore
@@ -9,6 +9,8 @@ lerna-debug.log*
 
 node_modules
 package-lock.json
+/dist
+build/
 dist
 dist-ssr
 *.local
