+++ src/App.tsx
@@ -1,6 +1,6 @@
 import './App.css'
 import { routes } from './routes';
-import { StoreProvider } from './zustand/store'
+
 import { RouterProvider } from 'react-router-dom';
 
 
@@ -8,9 +8,9 @@ function App() {
 
 
   return (
-    <StoreProvider>
+    <>
       <RouterProvider router={routes}/>
-    </StoreProvider>
+    </>
   )
 }
