+++ vite.config.ts
@@ -4,4 +4,13 @@ import react from '@vitejs/plugin-react-swc'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  build:{
+    target:'esnext',
+    outDir:'dist',
+    rollupOptions:{
+      output:{
+        manualChunks:undefined,
+      }
+    }
+  }
 })
